Create a friends-list page & extract this profile component
